Remove usage of doubles for vectors

+++ detectors/pcl_detector/src/optics_detector.cpp
@@ -14,7 +14,7 @@ pcl::PointCloud<pcl::PointXYZ> OPTICSDetector::get_detections(const pcl::PointCl
 
     std::vector<bool> visited(points.size(), false);
     std::vector<int> predecessor(points.size(), -1);
-    std::vector<double> reachability_distance(points.size(), std::numeric_limits<double>::infinity());
+    std::vector<float> reachability_distance(points.size(), std::numeric_limits<float>::infinity());
 
     for (size_t i = 0; i < points.size(); ++i) {
         if (!visited[i]) {
@@ -58,14 +58,14 @@ void OPTICSDetector::expand_cluster(const pcl::PointCloud<pcl::PointXYZ>& points
                     int point_idx,
                     std::vector<int>& predecessor,
                     std::vector<bool>& visited,
-                    std::vector<double>& reachability_distance) {
+                    std::vector<float>& reachability_distance) {
     visited[point_idx] = true;
 
     std::vector<int> neighbors;
     std::vector<float> distances;
     kdtree.radiusSearch(point_idx, eps_, neighbors, distances);
 
-    std::vector<double> core_distances(neighbors.size());
+    std::vector<float> core_distances(neighbors.size());
     for (size_t i = 0; i < neighbors.size(); ++i) {
         int neighbor_idx = neighbors[i];
         if (!visited[neighbor_idx]) {
@@ -76,15 +76,15 @@ void OPTICSDetector::expand_cluster(const pcl::PointCloud<pcl::PointXYZ>& points
             kdtree.radiusSearch(neighbor_idx, eps_, neighbor_neighbors, neighbor_distances);
 
             if (neighbor_neighbors.size() >= min_pts_) {
-                std::vector<double> neighbor_core_distances(neighbor_neighbors.size());
+                std::vector<float> neighbor_core_distances(neighbor_neighbors.size());
                 for (size_t j = 0; j < neighbor_neighbors.size(); ++j) {
                     neighbor_core_distances[j] = pcl::euclideanDistance(points[neighbor_idx], points[neighbor_neighbors[j]]);
                 }
                 std::sort(neighbor_core_distances.begin(), neighbor_core_distances.end());
-                double core_distance = neighbor_core_distances[min_pts_ - 1];
+                float core_distance = neighbor_core_distances[min_pts_ - 1];
                 core_distances[i] = core_distance;
             } else {
-                core_distances[i] = std::numeric_limits<double>::infinity();
+                core_distances[i] = std::numeric_limits<float>::infinity();
             }
         } else {
             core_distances[i] = reachability_distance[neighbor_idx];
